fix: empty/not-found responses refactor: move deserializer to serialiser class
